ws/requirements.txt
awsebcli==3.7.6
blessed==1.9.5
botocore==1.4.21
cement==2.8.2
colorama==0.3.7
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.2
django-braces==1.9.0
django-cors-headers==1.1.0
django-material==0.8.0
django-oauth-toolkit==0.10.0
django-redis-cache==1.6.5
django-social-auth==0.7.28
djangorestframework==3.3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
gunicorn==19.6.0
httplib2==0.9.2
jmespath==0.9.0
mysqlclient==1.3.7
oauth2==1.9.0.post1
oauthlib==1.0.3
pathspec==0.3.4
Pillow==3.2.0
psycopg2==2.6.1
PyJWT==1.4.0
python-dateutil==2.5.3
PyYAML==3.11
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
semantic-version==2.5.0
six==1.10.0
texttable==0.8.4
wcwidth==0.1.6
websocket-client==0.37.0
whitenoise==3.2